Fixed editing features to have full proper saving / Need to add updating to gui on save modified:   export/gui.png modified:   export/newtest.png new file:   export/test.png modified:   main.py modified:   readme.md new file:   src/__pycache__/date_select.cpython-311.pyc modified:   src/__pycache__/event_gui.cpython-311.pyc deleted:    src/event_gui.py modified:   timelines/gui.json modified:   timelines/newtest.json
