+++ timelines/newtest.json
@@ -1,13 +1,13 @@
 {
-    "timeline_name": "test new",
+    "timeline_name": "newtest",
     "start_date": [
         6,
-        28,
-        2019
+        30,
+        2024
     ],
     "end_date": [
         6,
-        28,
+        30,
         2024
     ],
     "events": []
